move String() method to a separate file
